Rename misc/dicts.py to config_dicts.py
